Split non-display parts of line plot/graph into LinePlotDisplay.py.
