+++ nion/swift/LineGraphCanvasItem.py
@@ -28,6 +28,7 @@ from nion.swift import MimeTypes
 from nion.swift import Undo
 from nion.swift.model import DisplayItem
 from nion.swift.model import DocumentModel
+from nion.swift.model import LinePlotDisplay
 from nion.swift.model import UISettings
 from nion.swift.model import Utility
 from nion.ui import CanvasItem
@@ -39,360 +40,7 @@ from nion.utils import Geometry
 _NDArray = numpy.typing.NDArray[typing.Any]
 
 
-@dataclasses.dataclass
-class RegionInfo:
-    channels: typing.Tuple[float, float]
-    selected: bool
-    index: int
-    left_text: str
-    right_text: str
-    middle_text: str
-    label: typing.Optional[str]
-    style: typing.Optional[str]
-    color: typing.Optional[str]
-
-
 _ = gettext.gettext
-
-
-def nice_label(value: float, precision: int) -> str:
-    if math.trunc(math.log10(abs(value) + numpy.nextafter(0, 1))) > 4:
-        return (u"{0:0." + u"{0:d}".format(precision) + "e}").format(value)
-    else:
-        return (u"{0:0." + u"{0:d}".format(precision) + "f}").format(value)
-
-
-class AxisScale(typing.Protocol):
-
-    @property
-    def axis_scale_id(self) -> str: ...
-
-    def make_ticker(self, display_min: float, display_max: float) -> Geometry.Ticker: ...
-
-    def convert_calibrated_value_to_scaled_value(self, calibrated_value: float) -> float: ...
-
-    def convert_scaled_value_to_calibrated_value(self, scaled_value: float) -> float: ...
-
-    def convert_calibrated_array_to_scaled_array(self, calibrated_xdata: DataAndMetadata.DataAndMetadata) -> DataAndMetadata.DataAndMetadata | None: ...
-
-    def display_origin(self, display_min: float, display_max: float) -> float: ...
-
-    def adjust_calibrated_limits(self, calibrated_min: float | None, calibrated_max: float | None, min_specified: bool, max_specified: bool) -> tuple[float, float]: ...
-
-
-class LinearAxisScale(AxisScale):
-    axis_scale_id = "linear"
-
-    def make_ticker(self, display_min: float, display_max: float) -> Geometry.Ticker:
-        return Geometry.LinearTicker(display_min, display_max)
-
-    def convert_calibrated_value_to_scaled_value(self, calibrated_value: float) -> float:
-        return calibrated_value
-
-    def convert_scaled_value_to_calibrated_value(self, scaled_value: float) -> float:
-        return scaled_value
-
-    def convert_calibrated_array_to_scaled_array(self, calibrated_xdata: DataAndMetadata.DataAndMetadata) -> DataAndMetadata.DataAndMetadata | None:
-        return calibrated_xdata
-
-    def display_origin(self, display_min: float, display_max: float) -> float:
-        if display_min <= 0.0 <= display_max:
-            return 0.0
-        return display_min if 0.0 < display_min else display_max
-
-    def adjust_calibrated_limits(self, calibrated_min: float | None, calibrated_max: float | None, min_specified: bool, max_specified: bool) -> tuple[float, float]:
-        if (calibrated_min is None) or (not min_specified and calibrated_min > 0.0):
-            calibrated_min = 0.0
-        if (calibrated_max is None) or (not max_specified and calibrated_max < 0.0):
-            calibrated_max = 0.0
-        return calibrated_min, calibrated_max
-
-
-class LogAxisScale(AxisScale):
-    axis_scale_id = "log"
-
-    def make_ticker(self, display_min: float, display_max: float) -> Geometry.Ticker:
-        return Geometry.LogTicker(display_min, display_max)
-
-    def convert_calibrated_value_to_scaled_value(self, calibrated_value: float) -> float:
-        return math.log10(calibrated_value) if calibrated_value > 0 else float("-inf")
-
-    def convert_scaled_value_to_calibrated_value(self, scaled_value: float) -> float:
-        return math.pow(10, scaled_value)
-
-    def convert_calibrated_array_to_scaled_array(self, calibrated_xdata: DataAndMetadata.DataAndMetadata) -> DataAndMetadata.DataAndMetadata | None:
-        scaled_data = numpy.log10(numpy.where(calibrated_xdata.data <= 0, numpy.nan, calibrated_xdata.data.astype(float)))
-        return DataAndMetadata.new_data_and_metadata(
-            scaled_data,
-            intensity_calibration=Calibration.Calibration(units=calibrated_xdata.intensity_calibration.units),
-            dimensional_calibrations=calibrated_xdata.dimensional_calibrations
-        )
-
-    def display_origin(self, display_min: float, display_max: float) -> float:
-        return display_min
-
-    def adjust_calibrated_limits(self, calibrated_min: float | None, calibrated_max: float | None, min_specified: bool, max_specified: bool) -> tuple[float, float]:
-        min_val = calibrated_min if calibrated_min is not None else 0.0
-        max_val = calibrated_max if calibrated_max is not None else 0.0
-        return min_val, max_val
-
-
-_axis_scale_types: dict[str, AxisScale] = {
-    "linear": LinearAxisScale(),
-    "log": LogAxisScale()
-}
-
-def _get_axis_scale(axis_scale_id: str | None) -> AxisScale:
-    # return the axis scale for the optional axis_scale_id, or the linear style if not found
-    axis_scale = _axis_scale_types.get(axis_scale_id, None) if axis_scale_id else None
-    return axis_scale if axis_scale else LinearAxisScale()
-
-
-def calculate_scaled_xdata(xdata: DataAndMetadata.DataAndMetadata, axis_scale: AxisScale) -> DataAndMetadata.DataAndMetadata | None:
-    """Calculate the 'scaled xdata' for the given xdata and axis scale.
-
-    The 'scaled xdata' is the xdata (with a calibration) but with a new intensity calibration where origin=0 and scale=1.
-    """
-    scalar_uncalibrated_data = xdata.data if not xdata.is_data_rgb_type else Image.convert_to_grayscale(xdata.data)
-    calibrated_data = xdata.intensity_calibration.convert_array_to_calibrated_value(scalar_uncalibrated_data)
-    calibrated_xdata = DataAndMetadata.new_data_and_metadata(
-        calibrated_data,
-        intensity_calibration=Calibration.Calibration(units=xdata.intensity_calibration.units),
-        dimensional_calibrations=xdata.dimensional_calibrations
-    )
-    return axis_scale.convert_calibrated_array_to_scaled_array(calibrated_xdata)
-
-
-def calculate_y_axis(xdata_list: typing.Sequence[DataAndMetadata.DataAndMetadata | None], data_min: float | None, data_max: float | None, axis_scale_id: str | None) -> tuple[float, float, Geometry.Ticker]:
-    """Calculate the calibrated min/max and y-axis ticker for list of xdata.
-
-    xdata_list is the original calibrated data
-    data_min and data_max are calibrated values
-
-    Returns scaled_data_min, scaled_data_max, ticker
-    """
-    min_specified = data_min is not None
-    max_specified = data_max is not None
-
-    scaled_data_min_opt: float | None = None
-    scaled_data_max_opt: float | None = None
-
-    axis_scale = _get_axis_scale(axis_scale_id)
-
-    # Determine min/max in calibrated data space before converting to display space
-    for xdata in xdata_list:
-        if xdata and xdata.data_shape[-1] > 0:
-            scaled_xdata = calculate_scaled_xdata(xdata, axis_scale)
-            if scaled_xdata is not None:
-                scaled_data = scaled_xdata.data if numpy.issubdtype(scaled_xdata.data.dtype, numpy.floating) else scaled_xdata.data.astype(float)
-                if scaled_data is not None and scaled_data.size > 0:
-                    with warnings.catch_warnings():
-                        warnings.simplefilter("ignore", category=RuntimeWarning)
-                        scaled_data_min = float(numpy.nanmin(scaled_data))
-                        scaled_data_max = float(numpy.nanmax(scaled_data))
-                    if numpy.isfinite(scaled_data_min):
-                        scaled_data_min_opt = scaled_data_min if scaled_data_min_opt is None else min(scaled_data_min_opt, scaled_data_min)
-                    if numpy.isfinite(scaled_data_max):
-                        scaled_data_max_opt = scaled_data_max if scaled_data_max_opt is None else max(scaled_data_max_opt, scaled_data_max)
-
-    if min_specified:
-        calibrated_min = typing.cast(float, data_min)
-    else:
-        if scaled_data_min_opt is not None:
-            calibrated_min = axis_scale.convert_scaled_value_to_calibrated_value(scaled_data_min_opt)
-        else:
-            calibrated_min = None
-
-    if max_specified:
-        calibrated_max = typing.cast(float, data_max)
-    else:
-        if scaled_data_max_opt is not None:
-            calibrated_max = axis_scale.convert_scaled_value_to_calibrated_value(scaled_data_max_opt)
-        else:
-            calibrated_max = None
-
-    adjusted_calibrated_min, adjusted_calibrated_max = axis_scale.adjust_calibrated_limits(calibrated_min, calibrated_max, min_specified, max_specified)
-
-    # Convert calibrated limits to display space
-    scaled_data_min = axis_scale.convert_calibrated_value_to_scaled_value(adjusted_calibrated_min)
-    scaled_data_max = axis_scale.convert_calibrated_value_to_scaled_value(adjusted_calibrated_max)
-    if math.isnan(scaled_data_min) or math.isinf(scaled_data_min):
-        scaled_data_min = 0.0
-    if math.isnan(scaled_data_max) or math.isinf(scaled_data_max):
-        scaled_data_max = 0.0
-    scaled_data_min, scaled_data_max = min(scaled_data_min, scaled_data_max), max(scaled_data_min, scaled_data_max)
-    if scaled_data_min == scaled_data_max:
-        scaled_data_min -= 1.0
-        scaled_data_max += 1.0
-
-    ticker = axis_scale.make_ticker(scaled_data_min, scaled_data_max)
-    if not min_specified:
-        scaled_data_min = ticker.minimum
-    if not max_specified:
-        scaled_data_max = ticker.maximum
-
-    return scaled_data_min, scaled_data_max, ticker
-
-
-@dataclasses.dataclass
-class LegendEntry:
-    label: str
-    fill_color: typing.Optional[str]
-    stroke_color: typing.Optional[str]
-
-
-class LineGraphAxes:
-    """Track information about line graph axes."""
-
-    def __init__(self, data_scale: float, scaled_data_min: float, scaled_data_max: float, data_left: int,
-                 data_right: int, x_calibration: typing.Optional[Calibration.Calibration],
-                 y_calibration: typing.Optional[Calibration.Calibration], axis_scale_id: typing.Optional[str],
-                 y_ticker: Geometry.Ticker) -> None:
-        assert x_calibration is None or x_calibration.is_valid
-        assert y_calibration is None or y_calibration.is_valid
-        # these items are considered to be input items
-        self.data_scale = data_scale
-        self.__uncalibrated_left_channel = data_left
-        self.__uncalibrated_right_channel = data_right
-        self.x_calibration = x_calibration
-        self.y_calibration = y_calibration
-        self.axis_scale = _get_axis_scale(axis_scale_id)
-        self.__scaled_data_min = scaled_data_min
-        self.__scaled_data_max = scaled_data_max
-        self.__y_ticker = y_ticker
-
-    def __repr__(self) -> str:
-        return f"Axes {self.drawn_left_channel}:{self.drawn_right_channel} [{self.scaled_data_min},{self.scaled_data_max} {self.x_calibration} {self.y_calibration} {self.axis_scale} {self.data_scale}"
-
-    def __eq__(self, other: typing.Any) -> bool:
-        if not isinstance(other, LineGraphAxes):
-            return False
-        return (self.data_scale == other.data_scale and
-                self.__uncalibrated_left_channel == other.__uncalibrated_left_channel and
-                self.__uncalibrated_right_channel == other.__uncalibrated_right_channel and
-                self.x_calibration == other.x_calibration and
-                self.y_calibration == other.y_calibration and
-                self.axis_scale.axis_scale_id == other.axis_scale.axis_scale_id and
-                self.__scaled_data_min == other.__scaled_data_min and
-                self.__scaled_data_max == other.__scaled_data_max and
-                self.__y_ticker == other.__y_ticker)
-
-    def __hash__(self) -> int:
-        return hash((self.data_scale, self.__uncalibrated_left_channel, self.__uncalibrated_right_channel, self.x_calibration, self.y_calibration, self.axis_scale.axis_scale_id, self.__scaled_data_min, self.__scaled_data_max, self.__y_ticker))
-
-    @property
-    def y_ticker(self) -> Geometry.Ticker:
-        return self.__y_ticker
-
-    @property
-    def uncalibrated_data_min(self) -> float:
-        y_calibration = self.y_calibration if self.y_calibration else Calibration.Calibration()
-        calibrated_value_min = self.axis_scale.convert_scaled_value_to_calibrated_value(self.scaled_data_min)
-        return y_calibration.convert_from_calibrated_value(calibrated_value_min)
-
-    @property
-    def uncalibrated_data_max(self) -> float:
-        y_calibration = self.y_calibration if self.y_calibration else Calibration.Calibration()
-        calibrated_value_max = self.axis_scale.convert_scaled_value_to_calibrated_value(self.scaled_data_max)
-        return y_calibration.convert_from_calibrated_value(calibrated_value_max)
-
-    @property
-    def scaled_data_min(self) -> float:
-        return self.__scaled_data_min
-
-    @property
-    def scaled_data_max(self) -> float:
-        return self.__scaled_data_max
-
-    @property
-    def calibrated_value_max(self) -> float:
-        calibrated_value = self.axis_scale.convert_scaled_value_to_calibrated_value(self.__scaled_data_max)
-        return calibrated_value
-
-    @property
-    def calibrated_value_min(self) -> float:
-        calibrated_value = self.axis_scale.convert_scaled_value_to_calibrated_value(self.__scaled_data_min)
-        return calibrated_value
-
-    @property
-    def drawn_left_channel(self) -> int:
-        return self.__uncalibrated_left_channel
-
-    @property
-    def drawn_right_channel(self) -> int:
-        return self.__uncalibrated_right_channel
-
-    @property
-    def calibrated_left_channel(self) -> float:
-        return self.x_calibration.convert_to_calibrated_value(self.drawn_left_channel) if self.x_calibration else float(self.__uncalibrated_left_channel)
-
-    @property
-    def calibrated_right_channel(self) -> float:
-        return self.x_calibration.convert_to_calibrated_value(self.drawn_right_channel) if self.x_calibration else float(self.__uncalibrated_right_channel)
-
-    def calculate_y_ticks(self, plot_height: int, flag_minor: bool = False) -> typing.Sequence[typing.Tuple[float, str, bool]]:
-        """Calculate the y-axis items dependent on the plot height."""
-
-        calibrated_data_min = self.scaled_data_min
-        calibrated_data_max = self.scaled_data_max
-        calibrated_data_range = calibrated_data_max - calibrated_data_min
-
-        ticker = self.y_ticker
-        y_ticks: typing.List[typing.Tuple[float, str, bool]] = list()
-
-        for i, (tick_value, tick_label) in enumerate(zip(ticker.values, ticker.labels)):
-            if calibrated_data_range != 0.0:
-                y_tick = plot_height - plot_height * (tick_value - calibrated_data_min) / calibrated_data_range
-            else:
-                y_tick = plot_height - plot_height * 0.5
-            if y_tick >= 0 and y_tick <= plot_height:
-                if flag_minor:
-                    y_ticks.append((y_tick, tick_label, i in ticker.minor_tick_indices))
-                else:
-                    y_ticks.append((y_tick, tick_label, False))
-
-        return y_ticks
-
-    def convert_scaled_y_value_to_uncalibrated_value(self, scaled_y_value: float) -> float:
-        calibrated_y_value = self.axis_scale.convert_scaled_value_to_calibrated_value(scaled_y_value)
-        y_calibration = self.y_calibration
-        if y_calibration:
-            return y_calibration.convert_from_calibrated_value(calibrated_y_value)
-        return calibrated_y_value
-
-    def calculate_x_ticks(self, plot_width: int) -> typing.Sequence[typing.Tuple[float, str]]:
-        """Calculate the x-axis items dependent on the plot width."""
-
-        x_calibration = self.x_calibration
-
-        uncalibrated_data_left = self.__uncalibrated_left_channel
-        uncalibrated_data_right = self.__uncalibrated_right_channel
-
-        calibrated_data_left = x_calibration.convert_to_calibrated_value(uncalibrated_data_left) if x_calibration is not None else uncalibrated_data_left
-        calibrated_data_right = x_calibration.convert_to_calibrated_value(uncalibrated_data_right) if x_calibration is not None else uncalibrated_data_right
-        calibrated_data_left, calibrated_data_right = min(calibrated_data_left, calibrated_data_right), max(calibrated_data_left, calibrated_data_right)
-
-        graph_left, graph_right, tick_values, division, precision = Geometry.make_pretty_range(calibrated_data_left, calibrated_data_right)
-
-        drawn_data_width = self.drawn_right_channel - self.drawn_left_channel
-
-        x_ticks: typing.List[typing.Tuple[float, str]] = list()
-        if drawn_data_width > 0.0:
-            for tick_value in tick_values:
-                label = nice_label(tick_value, precision)
-                data_tick = x_calibration.convert_from_calibrated_value(tick_value) if x_calibration else tick_value
-                x_tick = plot_width * (data_tick - self.drawn_left_channel) / drawn_data_width
-                if x_tick >= 0 and x_tick <= plot_width:
-                    x_ticks.append((x_tick, label))
-
-        return x_ticks
-
-    def convert_calibrated_array_to_scaled_array(self, xdata: DataAndMetadata.DataAndMetadata) -> typing.Optional[DataAndMetadata.DataAndMetadata]:
-        """Calculate the 'calibrated xdata'.
-
-         The 'calibrated xdata' is the xdata (with a calibration) but with a new intensity calibration where origin=0 and scale=1.
-         """
-        return calculate_scaled_xdata(xdata, self.axis_scale)
 
 
 def draw_background(drawing_context: DrawingContext.DrawingContext, plot_rect: Geometry.IntRect, background_color: typing.Optional[typing.Union[str, DrawingContext.LinearGradient]]) -> None:
@@ -491,7 +139,7 @@ def calculate_line_graph(plot_height: int, plot_width: int, plot_origin_y: int, 
                          scaled_data_range: float, calibrated_left_channel: float, calibrated_right_channel: float,
                          line_graph_x_calibration: Calibration.Calibration,
                          rebin_cache: typing.Optional[typing.Dict[str, typing.Any]],
-                         axis_scale: AxisScale) -> typing.Tuple[typing.List[LineGraphSegment], int]:
+                         axis_scale: LinePlotDisplay.AxisScale) -> typing.Tuple[typing.List[LineGraphSegment], int]:
     """Calculate the line graph segments for the given parameters.
 
     Note: line_graph_x_calibration may not match the calibration of scaled_xdata since the x_calibration
@@ -624,7 +272,7 @@ def draw_marker(drawing_context: DrawingContext.DrawingContext, p: Geometry.Floa
 
 
 class LineGraphBackgroundCanvasItemComposer(CanvasItem.BaseComposer):
-    def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache, axes: typing.Optional[LineGraphAxes], draw_grid: bool, background_color: typing.Optional[typing.Union[str, DrawingContext.LinearGradient]]) -> None:
+    def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache, axes: typing.Optional[LinePlotDisplay.LineGraphAxes], draw_grid: bool, background_color: typing.Optional[typing.Union[str, DrawingContext.LinearGradient]]) -> None:
         super().__init__(canvas_item, layout_sizing, cache)
         self.__axes = axes
         self.__draw_grid = draw_grid
@@ -656,14 +304,14 @@ class LineGraphBackgroundCanvasItem(CanvasItem.AbstractCanvasItem):
 
     def __init__(self) -> None:
         super().__init__()
-        self.__axes: typing.Optional[LineGraphAxes] = None
+        self.__axes: typing.Optional[LinePlotDisplay.LineGraphAxes] = None
         self.draw_grid = True
         self.background_color = "#FFF"
 
     def _get_composer(self, composer_cache: CanvasItem.ComposerCache) -> CanvasItem.BaseComposer:
         return LineGraphBackgroundCanvasItemComposer(self, self.layout_sizing, composer_cache, self.__axes, self.draw_grid, self.background_color)
 
-    def set_axes(self, axes: typing.Optional[LineGraphAxes]) -> None:
+    def set_axes(self, axes: typing.Optional[LinePlotDisplay.LineGraphAxes]) -> None:
         if self.__axes != axes:
             self.__axes = axes
             self.update()
@@ -672,9 +320,9 @@ class LineGraphBackgroundCanvasItem(CanvasItem.AbstractCanvasItem):
 @dataclasses.dataclass(frozen=True)
 class MappedCalibratedDataAndMetadataCacheItem:
     xdata: typing.Optional[DataAndMetadata.DataAndMetadata]
-    axes: typing.Optional[LineGraphAxes]
+    axes: typing.Optional[LinePlotDisplay.LineGraphAxes]
 
-    def key(self) -> typing.Tuple[typing.Optional[int], typing.Optional[LineGraphAxes]]:
+    def key(self) -> typing.Tuple[typing.Optional[int], typing.Optional[LinePlotDisplay.LineGraphAxes]]:
         return id(self.xdata.data) if self.xdata else None, self.axes
 
     def calculate(self) -> typing.Optional[DataAndMetadata.DataAndMetadata]:
@@ -684,10 +332,10 @@ class MappedCalibratedDataAndMetadataCacheItem:
 @dataclasses.dataclass(frozen=True)
 class SegmentsCacheItem:
     scaled_xdata: typing.Optional[DataAndMetadata.DataAndMetadata]
-    axes: typing.Optional[LineGraphAxes]
+    axes: typing.Optional[LinePlotDisplay.LineGraphAxes]
     canvas_bounds: Geometry.IntRect
 
-    def key(self) -> typing.Tuple[typing.Optional[int], typing.Optional[LineGraphAxes], Geometry.IntRect]:
+    def key(self) -> typing.Tuple[typing.Optional[int], typing.Optional[LinePlotDisplay.LineGraphAxes], Geometry.IntRect]:
         return id(self.scaled_xdata.data) if self.scaled_xdata else None, self.axes, self.canvas_bounds
 
     def calculate(self) -> typing.Tuple[typing.List[LineGraphSegment], float]:
@@ -723,90 +371,36 @@ class SegmentsCacheItem:
         return segments, baseline
 
 
-class LineGraphLayer:
-    """Represents a layer of the line graph.
+def draw_fills(line_graph_layer: LinePlotDisplay.LineGraphLayer, drawing_context: DrawingContext.DrawingContext, canvas_bounds: Geometry.IntRect, composer_cache: CanvasItem.ComposerCache) -> typing.Tuple[CanvasItem.CacheValue, ...]:
+    if line_graph_layer.fill_color:
+        scaled_data_and_metadata_cache_item = MappedCalibratedDataAndMetadataCacheItem(line_graph_layer.xdata, line_graph_layer.axes)
+        scaled_xdata_cache_value = composer_cache.get_cache_value(scaled_data_and_metadata_cache_item)
+        scaled_xdata = typing.cast(typing.Optional[DataAndMetadata.DataAndMetadata], scaled_xdata_cache_value.value)
+        segments_cache_item = SegmentsCacheItem(scaled_xdata, line_graph_layer.axes, canvas_bounds)
+        segments_cache_value = composer_cache.get_cache_value(segments_cache_item)
+        segments, baseline = typing.cast(typing.Tuple[typing.List[LineGraphSegment], float], segments_cache_value.value)
+        for segment in segments:
+            segment.fill(drawing_context, baseline, Color.Color(line_graph_layer.fill_color))
+        return scaled_xdata_cache_value, segments_cache_value
+    return tuple()
 
-    xdata is calibrated data.
 
-    Tracks the data, calibrated data, axes. Provides methods to calculate the segments and draw fills/strokes separately.
-    """
-
-    def __init__(self,
-                 xdata: typing.Optional[DataAndMetadata.DataAndMetadata],
-                 axes: typing.Optional[LineGraphAxes],
-                 fill_color: typing.Optional[Color.Color],
-                 stroke_color: typing.Optional[Color.Color],
-                 stroke_width: typing.Optional[float]) -> None:
-        self.__xdata = xdata
-        self.__fill_color = fill_color
-        self.__stroke_color = stroke_color
-        self.__stroke_width = stroke_width or 0.5
-        self.__canvas_bounds: typing.Optional[Geometry.IntRect] = None
-        self.__calibrated_xdata: typing.Optional[DataAndMetadata.DataAndMetadata] = None
-        self.__axes = axes
-
-    def __repr__(self) -> str:
-        data_sum = numpy.sum(self.__xdata) if self.__xdata else 0.0
-        return f"LineGraphLayer {data_sum} {self.__fill_color} {self.__stroke_color} {self.__stroke_width} {self.__axes}"
-
-    def __eq__(self, other: typing.Any) -> bool:
-        if not isinstance(other, LineGraphLayer):
-            return False
-        return (self.__xdata is not None and other.__xdata is not None and self.__xdata.data is other.__xdata.data and
-                self.__fill_color == other.__fill_color and
-                self.__stroke_color == other.__stroke_color and
-                self.__stroke_width == other.__stroke_width and
-                self.__axes == other.__axes)
-
-    @property
-    def xdata(self) -> typing.Optional[DataAndMetadata.DataAndMetadata]:
-        return self.__xdata
-
-    @property
-    def fill_color(self) -> typing.Optional[Color.Color]:
-        return self.__fill_color
-
-    @property
-    def stroke_color(self) -> typing.Optional[Color.Color]:
-        return self.__stroke_color
-
-    @property
-    def stroke_width(self) -> float:
-        return self.__stroke_width
-
-    @property
-    def axes(self) -> typing.Optional[LineGraphAxes]:
-        return self.__axes
-
-    def draw_fills(self, drawing_context: DrawingContext.DrawingContext, canvas_bounds: Geometry.IntRect, composer_cache: CanvasItem.ComposerCache) -> typing.Tuple[CanvasItem.CacheValue, ...]:
-        if self.fill_color:
-            scaled_data_and_metadata_cache_item = MappedCalibratedDataAndMetadataCacheItem(self.__xdata, self.__axes)
-            scaled_xdata_cache_value = composer_cache.get_cache_value(scaled_data_and_metadata_cache_item)
-            scaled_xdata = typing.cast(typing.Optional[DataAndMetadata.DataAndMetadata], scaled_xdata_cache_value.value)
-            segments_cache_item = SegmentsCacheItem(scaled_xdata, self.__axes, canvas_bounds)
-            segments_cache_value = composer_cache.get_cache_value(segments_cache_item)
-            segments, baseline = typing.cast(typing.Tuple[typing.List[LineGraphSegment], float], segments_cache_value.value)
-            for segment in segments:
-                segment.fill(drawing_context, baseline, Color.Color(self.fill_color))
-            return scaled_xdata_cache_value, segments_cache_value
-        return tuple()
-
-    def draw_strokes(self, drawing_context: DrawingContext.DrawingContext, canvas_bounds: Geometry.IntRect, composer_cache: CanvasItem.ComposerCache) -> typing.Tuple[CanvasItem.CacheValue, ...]:
-        if self.stroke_color:
-            scaled_data_and_metadata_cache_item = MappedCalibratedDataAndMetadataCacheItem(self.__xdata, self.__axes)
-            scaled_xdata_cache_value = composer_cache.get_cache_value(scaled_data_and_metadata_cache_item)
-            scaled_xdata = typing.cast(typing.Optional[DataAndMetadata.DataAndMetadata], scaled_xdata_cache_value.value)
-            segments_cache_item = SegmentsCacheItem(scaled_xdata, self.__axes, canvas_bounds)
-            segments_cache_value = composer_cache.get_cache_value(segments_cache_item)
-            segments, baseline = typing.cast(typing.Tuple[typing.List[LineGraphSegment], float], segments_cache_value.value)
-            for segment in segments:
-                segment.stroke(drawing_context, baseline, Color.Color(self.stroke_color), self.stroke_width)
-            return scaled_xdata_cache_value, segments_cache_value
-        return tuple()
+def draw_strokes(line_graph_layer: LinePlotDisplay.LineGraphLayer, drawing_context: DrawingContext.DrawingContext, canvas_bounds: Geometry.IntRect, composer_cache: CanvasItem.ComposerCache) -> typing.Tuple[CanvasItem.CacheValue, ...]:
+    if line_graph_layer.stroke_color:
+        scaled_data_and_metadata_cache_item = MappedCalibratedDataAndMetadataCacheItem(line_graph_layer.xdata, line_graph_layer.axes)
+        scaled_xdata_cache_value = composer_cache.get_cache_value(scaled_data_and_metadata_cache_item)
+        scaled_xdata = typing.cast(typing.Optional[DataAndMetadata.DataAndMetadata], scaled_xdata_cache_value.value)
+        segments_cache_item = SegmentsCacheItem(scaled_xdata, line_graph_layer.axes, canvas_bounds)
+        segments_cache_value = composer_cache.get_cache_value(segments_cache_item)
+        segments, baseline = typing.cast(typing.Tuple[typing.List[LineGraphSegment], float], segments_cache_value.value)
+        for segment in segments:
+            segment.stroke(drawing_context, baseline, Color.Color(line_graph_layer.stroke_color), line_graph_layer.stroke_width)
+        return scaled_xdata_cache_value, segments_cache_value
+    return tuple()
 
 
 class LineGraphLayerCanvasItemComposer(CanvasItem.BaseComposer):
-    def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache, line_graph_layer: LineGraphLayer, is_fill: bool) -> None:
+    def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache, line_graph_layer: LinePlotDisplay.LineGraphLayer, is_fill: bool) -> None:
         super().__init__(canvas_item, layout_sizing, cache)
         self.__cache_values = list[typing.Tuple[CanvasItem.CacheValue, ...]]()
         self.__line_graph_layer = line_graph_layer
@@ -816,9 +410,9 @@ class LineGraphLayerCanvasItemComposer(CanvasItem.BaseComposer):
         cache_values = list[typing.Tuple[CanvasItem.CacheValue, ...]]()
         line_graph_layer = self.__line_graph_layer
         if self.__is_fill:
-            cache_values.append(line_graph_layer.draw_fills(drawing_context, canvas_bounds, composer_cache))
+            cache_values.append(draw_fills(line_graph_layer, drawing_context, canvas_bounds, composer_cache))
         else:
-            cache_values.append(line_graph_layer.draw_strokes(drawing_context, canvas_bounds, composer_cache))
+            cache_values.append(draw_strokes(line_graph_layer, drawing_context, canvas_bounds, composer_cache))
         self.__cache_values = cache_values
 
 
@@ -830,7 +424,7 @@ class LineGraphLayerCanvasItem(CanvasItem.AbstractCanvasItem):
 
     def __init__(self, cache: CanvasItem.ComposerCache) -> None:
         super().__init__(cache)
-        self.__line_graph_layer: typing.Optional[LineGraphLayer] = None
+        self.__line_graph_layer: typing.Optional[LinePlotDisplay.LineGraphLayer] = None
         self.__is_fill = False
 
     def _get_composer(self, composer_cache: CanvasItem.ComposerCache) -> CanvasItem.BaseComposer:
@@ -840,7 +434,7 @@ class LineGraphLayerCanvasItem(CanvasItem.AbstractCanvasItem):
         else:
             return CanvasItem.EmptyCanvasItemComposer(self, self.layout_sizing, composer_cache)
 
-    def update_line_graph_layer(self, line_graph_layer: LineGraphLayer, is_fill: bool) -> None:
+    def update_line_graph_layer(self, line_graph_layer: LinePlotDisplay.LineGraphLayer, is_fill: bool) -> None:
         if self.__line_graph_layer != line_graph_layer or self.__is_fill != is_fill:
             self.__line_graph_layer = line_graph_layer
             self.__is_fill = is_fill
@@ -851,7 +445,7 @@ class LineGraphLayerCanvasItem(CanvasItem.AbstractCanvasItem):
         return self.__line_graph_layer.xdata if self.__line_graph_layer else None
 
     @property
-    def _axes_for_testing(self) -> typing.Optional[LineGraphAxes]:
+    def _axes_for_testing(self) -> typing.Optional[LinePlotDisplay.LineGraphAxes]:
         return self.__line_graph_layer.axes if self.__line_graph_layer else None
 
     @property
@@ -898,7 +492,7 @@ class LineGraphLayersCanvasItem(CanvasItem.CanvasItemComposition):
         self.__display_frame_rate_id = value
         self.update()
 
-    def update_line_graph_layers(self, line_graph_layers: typing.Sequence[LineGraphLayer]) -> None:
+    def update_line_graph_layers(self, line_graph_layers: typing.Sequence[LinePlotDisplay.LineGraphLayer]) -> None:
         line_graph_layers = list(line_graph_layers)
         while len(self.canvas_items) < len(line_graph_layers) * 2:
             self.add_canvas_item(LineGraphLayerCanvasItem(self.__cache))
@@ -919,7 +513,7 @@ class LineGraphLayersCanvasItem(CanvasItem.CanvasItemComposition):
         return typing.cast(LineGraphLayerCanvasItem, self.canvas_items[0])._xdata_for_testing if self.canvas_items else None
 
     @property
-    def _axes_for_testing(self) -> typing.Optional[LineGraphAxes]:
+    def _axes_for_testing(self) -> typing.Optional[LinePlotDisplay.LineGraphAxes]:
         return typing.cast(LineGraphLayerCanvasItem, self.canvas_items[0])._axes_for_testing if self.canvas_items else None
 
     @property
@@ -928,7 +522,7 @@ class LineGraphLayersCanvasItem(CanvasItem.CanvasItemComposition):
 
 
 class LineGraphRegionsCanvasItemComposer(CanvasItem.BaseComposer):
-    def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache, axes: typing.Optional[LineGraphAxes], regions: typing.Sequence[RegionInfo], is_focused: bool, ui_settings: UISettings.UISettings) -> None:
+    def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache, axes: typing.Optional[LinePlotDisplay.LineGraphAxes], regions: typing.Sequence[LinePlotDisplay.RegionInfo], is_focused: bool, ui_settings: UISettings.UISettings) -> None:
         super().__init__(canvas_item, layout_sizing, cache)
         self.__cache_values = list[typing.Tuple[CanvasItem.CacheValue, ...]]()
         self.__axes = axes
@@ -1079,22 +673,22 @@ class LineGraphRegionsCanvasItem(CanvasItem.AbstractCanvasItem):
 
     def __init__(self, ui_settings: UISettings.UISettings) -> None:
         super().__init__()
-        self.__regions = tuple[RegionInfo]()
-        self.__axes: LineGraphAxes | None = None
+        self.__regions = tuple[LinePlotDisplay.RegionInfo]()
+        self.__axes: LinePlotDisplay.LineGraphAxes | None = None
         self.__is_focused = False
         self.__ui_settings = ui_settings
 
     @property
-    def _axes(self) -> LineGraphAxes | None:  # for testing only
+    def _axes(self) -> LinePlotDisplay.LineGraphAxes | None:  # for testing only
         return self.__axes
 
-    def update_line_graph_layers(self, line_graph_layers: typing.Sequence[LineGraphLayer]) -> None:
+    def update_line_graph_layers(self, line_graph_layers: typing.Sequence[LinePlotDisplay.LineGraphLayer]) -> None:
         axes = line_graph_layers[0].axes if line_graph_layers else None
         if self.__axes != axes:
             self.__axes = axes
             self.update()
 
-    def set_regions(self, regions: typing.Sequence[RegionInfo]) -> None:
+    def set_regions(self, regions: typing.Sequence[LinePlotDisplay.RegionInfo]) -> None:
         regions = tuple(regions)
         if (self.__regions is None and regions is not None) or (self.__regions != regions):
             self.__regions = regions
@@ -1136,7 +730,7 @@ class LineGraphFrameCanvasItem(CanvasItem.AbstractCanvasItem):
 
 
 class LineGraphHorizontalAxisTicksCanvasItemComposer(CanvasItem.BaseComposer):
-    def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache, axes: typing.Optional[LineGraphAxes], tick_height: int) -> None:
+    def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache, axes: typing.Optional[LinePlotDisplay.LineGraphAxes], tick_height: int) -> None:
         super().__init__(canvas_item, layout_sizing, cache)
         self.__axes = axes
         self.tick_height = tick_height
@@ -1166,11 +760,11 @@ class LineGraphHorizontalAxisTicksCanvasItem(CanvasItem.AbstractCanvasItem):
 
     def __init__(self) -> None:
         super().__init__()
-        self.__axes: typing.Optional[LineGraphAxes] = None
+        self.__axes: typing.Optional[LinePlotDisplay.LineGraphAxes] = None
         self.__tick_height = 4
         self.update_sizing(self.sizing.with_fixed_height(self.__tick_height))
 
-    def set_axes(self, axes: typing.Optional[LineGraphAxes]) -> None:
+    def set_axes(self, axes: typing.Optional[LinePlotDisplay.LineGraphAxes]) -> None:
         if self.__axes != axes:
             self.__axes = axes
             self.update()
@@ -1180,7 +774,7 @@ class LineGraphHorizontalAxisTicksCanvasItem(CanvasItem.AbstractCanvasItem):
 
 
 class LineGraphHorizontalAxisScaleCanvasItemComposer(CanvasItem.BaseComposer):
-    def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache, axes: typing.Optional[LineGraphAxes], font_size: int) -> None:
+    def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache, axes: typing.Optional[LinePlotDisplay.LineGraphAxes], font_size: int) -> None:
         super().__init__(canvas_item, layout_sizing, cache)
         self.__axes = axes
         self.__font_size = font_size
@@ -1208,11 +802,11 @@ class LineGraphHorizontalAxisScaleCanvasItem(CanvasItem.AbstractCanvasItem):
 
     def __init__(self) -> None:
         super().__init__()
-        self.__axes: typing.Optional[LineGraphAxes] = None
+        self.__axes: typing.Optional[LinePlotDisplay.LineGraphAxes] = None
         self.__font_size = 12
         self.update_sizing(self.sizing.with_fixed_height(self.__font_size + 4))
 
-    def set_axes(self, axes: typing.Optional[LineGraphAxes]) -> None:
+    def set_axes(self, axes: typing.Optional[LinePlotDisplay.LineGraphAxes]) -> None:
         if self.__axes != axes:
             self.__axes = axes
             self.update()
@@ -1222,7 +816,7 @@ class LineGraphHorizontalAxisScaleCanvasItem(CanvasItem.AbstractCanvasItem):
 
 
 class LineGraphHorizontalAxisLabelCanvasItemComposer(CanvasItem.BaseComposer):
-    def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache, axes: typing.Optional[LineGraphAxes], font_size: int) -> None:
+    def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache, axes: typing.Optional[LinePlotDisplay.LineGraphAxes], font_size: int) -> None:
         super().__init__(canvas_item, layout_sizing, cache)
         self.__axes = axes
         self.__font_size = font_size
@@ -1250,7 +844,7 @@ class LineGraphHorizontalAxisLabelCanvasItem(CanvasItem.AbstractCanvasItem):
 
     def __init__(self) -> None:
         super().__init__()
-        self.__axes: typing.Optional[LineGraphAxes] = None
+        self.__axes: typing.Optional[LinePlotDisplay.LineGraphAxes] = None
         self.__font_size = 12
         self.update_sizing(self.sizing.with_fixed_height(self.__font_size + 4))
 
@@ -1266,7 +860,7 @@ class LineGraphHorizontalAxisLabelCanvasItem(CanvasItem.AbstractCanvasItem):
                 new_sizing = new_sizing.with_maximum_height(self.__font_size + 4)
         self.update_sizing(new_sizing)
 
-    def set_axes(self, axes: typing.Optional[LineGraphAxes]) -> None:
+    def set_axes(self, axes: typing.Optional[LinePlotDisplay.LineGraphAxes]) -> None:
         if self.__axes != axes:
             self.__axes = axes
             self.size_to_content()
@@ -1277,7 +871,7 @@ class LineGraphHorizontalAxisLabelCanvasItem(CanvasItem.AbstractCanvasItem):
 
 
 class LineGraphVerticalAxisTicksCanvasItemComposer(CanvasItem.BaseComposer):
-    def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache, axes: typing.Optional[LineGraphAxes], tick_width: int) -> None:
+    def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache, axes: typing.Optional[LinePlotDisplay.LineGraphAxes], tick_width: int) -> None:
         super().__init__(canvas_item, layout_sizing, cache)
         self.__axes = axes
         self.__tick_width = tick_width
@@ -1306,11 +900,11 @@ class LineGraphVerticalAxisTicksCanvasItem(CanvasItem.AbstractCanvasItem):
 
     def __init__(self) -> None:
         super().__init__()
-        self.__axes: typing.Optional[LineGraphAxes] = None
+        self.__axes: typing.Optional[LinePlotDisplay.LineGraphAxes] = None
         self.__tick_width = 4
         self.update_sizing(self.sizing.with_fixed_width(self.__tick_width))
 
-    def set_axes(self, axes: typing.Optional[LineGraphAxes]) -> None:
+    def set_axes(self, axes: typing.Optional[LinePlotDisplay.LineGraphAxes]) -> None:
         if self.__axes != axes:
             self.__axes = axes
             self.update()
@@ -1375,7 +969,7 @@ def calculate_scientific_notation_drawing_width(ui_settings: UISettings.UISettin
 
 
 class LineGraphVerticalAxisScaleCanvasItemComposer(CanvasItem.BaseComposer):
-    def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache, axes: typing.Optional[LineGraphAxes], font_size: int, fonts: typing.Tuple[str, str], ui_settings: UISettings.UISettings) -> None:
+    def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache, axes: typing.Optional[LinePlotDisplay.LineGraphAxes], font_size: int, fonts: typing.Tuple[str, str], ui_settings: UISettings.UISettings) -> None:
         super().__init__(canvas_item, layout_sizing, cache)
         self.__axes = axes
         self.__font_size = font_size
@@ -1428,7 +1022,7 @@ class LineGraphVerticalAxisScaleCanvasItem(CanvasItem.AbstractCanvasItem):
 
     def __init__(self, ui_settings: UISettings.UISettings) -> None:
         super().__init__()
-        self.__axes: typing.Optional[LineGraphAxes] = None
+        self.__axes: typing.Optional[LinePlotDisplay.LineGraphAxes] = None
         self.__font_size = 12
         self.__fonts = ("{0:d}px".format(self.__font_size), "{0:d}px".format(int(self.__font_size * 0.8)))
         self.__ui_settings = ui_settings
@@ -1453,7 +1047,7 @@ class LineGraphVerticalAxisScaleCanvasItem(CanvasItem.AbstractCanvasItem):
 
         self.update_sizing(new_sizing)
 
-    def set_axes(self, axes: typing.Optional[LineGraphAxes]) -> None:
+    def set_axes(self, axes: typing.Optional[LinePlotDisplay.LineGraphAxes]) -> None:
         if self.__axes != axes:
             self.__axes = axes
             self.size_to_content()
@@ -1464,7 +1058,7 @@ class LineGraphVerticalAxisScaleCanvasItem(CanvasItem.AbstractCanvasItem):
 
 
 class LineGraphVerticalAxisLabelCanvasItemComposer(CanvasItem.BaseComposer):
-    def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache, axes: typing.Optional[LineGraphAxes], font_size: int) -> None:
+    def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache, axes: typing.Optional[LinePlotDisplay.LineGraphAxes], font_size: int) -> None:
         super().__init__(canvas_item, layout_sizing, cache)
         self.__axes = axes
         self.__font_size = font_size
@@ -1498,7 +1092,7 @@ class LineGraphVerticalAxisLabelCanvasItem(CanvasItem.AbstractCanvasItem):
 
     def __init__(self) -> None:
         super().__init__()
-        self.__axes: typing.Optional[LineGraphAxes] = None
+        self.__axes: typing.Optional[LinePlotDisplay.LineGraphAxes] = None
         self.__font_size = 12
         self.update_sizing(self.sizing.with_fixed_width(self.__font_size + 4))
 
@@ -1514,7 +1108,7 @@ class LineGraphVerticalAxisLabelCanvasItem(CanvasItem.AbstractCanvasItem):
                 new_sizing = new_sizing.with_maximum_width(self.__font_size + 4)
         self.update_sizing(new_sizing)
 
-    def set_axes(self, axes: typing.Optional[LineGraphAxes]) -> None:
+    def set_axes(self, axes: typing.Optional[LinePlotDisplay.LineGraphAxes]) -> None:
         if self.__axes != axes:
             self.__axes = axes
             self.size_to_content()
@@ -1545,8 +1139,8 @@ class LineGraphLegendCanvasItemDelegate(typing.Protocol):
 
 class LineGraphLegendCanvasItemComposer(CanvasItem.BaseComposer):
     def __init__(self, canvas_item: CanvasItem.AbstractCanvasItem, layout_sizing: CanvasItem.Sizing, cache: CanvasItem.ComposerCache,
-                 legend_entries: typing.Sequence[LegendEntry], effective_legend_entries: typing.Sequence[LegendEntry],
-                 foreign_legend_entry: typing.Optional[LegendEntry], mouse_pressed_for_dragging: bool,
+                 legend_entries: typing.Sequence[LinePlotDisplay.LegendEntry], effective_legend_entries: typing.Sequence[LinePlotDisplay.LegendEntry],
+                 foreign_legend_entry: typing.Optional[LinePlotDisplay.LegendEntry], mouse_pressed_for_dragging: bool,
                  entry_to_insert: typing.Optional[int], font_size: int, ui_settings: UISettings.UISettings) -> None:
         super().__init__(canvas_item, layout_sizing, cache)
         self.__legend_entries = list(legend_entries)
@@ -1636,11 +1230,11 @@ class LineGraphLegendCanvasItem(CanvasItem.AbstractCanvasItem):
         self.__delegate = delegate
 
         # current legend items corresponding to layers of the display item
-        self.__legend_entries: typing.List[LegendEntry] = list()
+        self.__legend_entries: typing.List[LinePlotDisplay.LegendEntry] = list()
 
         # reordered/inserted entries that have not yet been applied. also last for change detection purposes.
-        self.__effective_entries = list[LegendEntry]()
-        self.__last_effective_entries = list[LegendEntry]()
+        self.__effective_entries = list[LinePlotDisplay.LegendEntry]()
+        self.__last_effective_entries = list[LinePlotDisplay.LegendEntry]()
 
         # True when a user is clicking, but a drag hasn't started because they haven't moved the mouse enough
         self.__mouse_pressed_for_dragging = False
@@ -1658,7 +1252,7 @@ class LineGraphLegendCanvasItem(CanvasItem.AbstractCanvasItem):
         self.__entry_to_insert: typing.Optional[int] = None
 
         # An entry from another display item that we're dragging into this one
-        self.__foreign_legend_entry: typing.Optional[LegendEntry] = None
+        self.__foreign_legend_entry: typing.Optional[LinePlotDisplay.LegendEntry] = None
 
         # The index and UUID of the other display item layer we're dragging here
         self.__foreign_legend_uuid_and_index: typing.Optional[typing.Tuple[DisplayItem.DisplayItem, int]] = None
@@ -1719,7 +1313,7 @@ class LineGraphLegendCanvasItem(CanvasItem.AbstractCanvasItem):
     def wants_drag_event(self, mime_data: UserInterface.MimeData, x: int, y: int) -> bool:
         return mime_data.has_format(MimeTypes.LAYER_MIME_TYPE)
 
-    def set_legend_entries(self, legend_entries: typing.Sequence[LegendEntry]) -> None:
+    def set_legend_entries(self, legend_entries: typing.Sequence[LinePlotDisplay.LegendEntry]) -> None:
         # assume this is only called when legend entries changes
         self.__legend_entries = list(legend_entries)
         self.__generate_effective_entries()
@@ -1853,7 +1447,7 @@ class LineGraphLegendCanvasItem(CanvasItem.AbstractCanvasItem):
             else:
                 # if we aren't the source, setup foreign_legend_* and update the display
                 assert display_item
-                self.__foreign_legend_entry = LegendEntry(label=legend_data["label"], fill_color=legend_data["fill_color"], stroke_color=legend_data["stroke_color"])
+                self.__foreign_legend_entry = LinePlotDisplay.LegendEntry(label=legend_data["label"], fill_color=legend_data["fill_color"], stroke_color=legend_data["stroke_color"])
                 self.__foreign_legend_uuid_and_index = (display_item, typing.cast(int, legend_data["index"]))
                 self.update()
             return "move"
@@ -1894,7 +1488,7 @@ class LineGraphLegendCanvasItem(CanvasItem.AbstractCanvasItem):
         return "ignore"
 
     @property
-    def effective_entries(self) -> typing.Sequence[LegendEntry]:
+    def effective_entries(self) -> typing.Sequence[LinePlotDisplay.LegendEntry]:
         return self.__effective_entries
 
     def _get_composer(self, composer_cache: CanvasItem.ComposerCache) -> CanvasItem.BaseComposer:
